Deprecated due to a better alternative.

+++ src/main/java/io/github/thesummergrinch/mcmanhunt/cache/UniverseCache.java
@@ -46,9 +46,10 @@ public final class UniverseCache implements ConfigurationSerializable {
         this.universeCache.remove(universeName);
     }
 
+    @Deprecated
     public void onDisable() {
         universeCache.values().forEach(universe -> {
-            if (universe.getMarkedForDestruction()) {
+            if (universe.getMarkedForDestruction() && !universe.isInUse()) {
                 universe.destroyUniverse();
             }
         });
